+++ lib/cbor/cbor.dart
@@ -5,3 +5,5 @@ library cbor;
 export 'core/cbor.dart';
 
 export 'types/types.dart';
+
+export 'exception/exception.dart';