+++ lib/cbor/exception/exception.dart
@@ -1,0 +1,6 @@
+import 'package:blockchain_utils/exception/exceptions.dart';
+
+class CborException extends BlockchainUtilsException {
+  const CborException(String message, {Map<String, dynamic>? details})
+      : super(message, details: details);
+}