+++ lib/cbor/types/list.dart
@@ -1,4 +1,3 @@
-import 'package:blockchain_utils/helper/helper.dart';
 import 'package:blockchain_utils/utils/utils.dart';
 import 'package:blockchain_utils/cbor/utils/dynamic_bytes.dart';
 import 'package:blockchain_utils/cbor/core/tags.dart';
@@ -9,18 +8,13 @@ class CborListValue<T> implements CborObject {
   /// Constructor for creating a CborListValue instance with the provided parameters.
   /// It accepts the List of all cbor encodable value.
   ///
-  CborListValue.fixedLength(List<T> value)
-      : value = value.immutable,
-        _isFixedLength = true;
+  CborListValue.fixedLength(this.value) : _isFixedLength = true;
 
   /// Constructor for creating a CborListValue instance with the provided parameters.
   /// It accepts the List of all cbor encodable value.
   /// this method encode values with indefinite tag.
-  CborListValue.dynamicLength(List<T> value)
-      : value = value.immutable,
-        _isFixedLength = false;
+  CborListValue.dynamicLength(this.value) : _isFixedLength = false;
 
-  /// value as List
   @override
   final List<T> value;